add exception handling

+++ src/main/java/com/finance/services/AccountService.java
@@ -2,12 +2,15 @@ package com.finance.services;
 
 import com.finance.dto.AccountDTO;
 import com.finance.entity.Account;
+import com.finance.exceptions.DublicateErrorException;
+import com.finance.exceptions.RecordNotFoundException;
 import com.finance.mapper.AccountMapper;
 import com.finance.repo.AccountRepo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -24,8 +27,16 @@ public class AccountService {
 
 
     public AccountDTO insert(Account account) {
-        AccountDTO dto= accountMapper.mapToDTO(account);
+
+        if(!account.getEmail().isEmpty()&&account.getEmail()!=null){
+            Optional<Account> entity= accountRepo.findByEmail(account.getEmail());
+            if(entity.isPresent())
+                throw new DublicateErrorException("This email is already used ");
+
+        }
+        account.setDateTime(LocalDateTime.now());
         accountRepo.save(account);
+        AccountDTO dto= accountMapper.mapToDTO(account);
         return dto;
     }
 
@@ -33,7 +44,7 @@ public class AccountService {
     public AccountDTO update(Account account) {
         Optional<Account> check = accountRepo.findById(account.getID());
         if (check.isEmpty())
-            throw new IllegalArgumentException("Not found to update");
+            throw new RecordNotFoundException("Not found to update");
 
         AccountDTO dto = accountMapper.mapToDTO(account);
         accountRepo.save(account);
@@ -41,8 +52,14 @@ public class AccountService {
     }
 
     public AccountDTO findAccountByID(Long id) {
+        Optional<Account> entity =accountRepo.findById(id);
+        if(entity.isEmpty()&&entity==null)
+            throw new RecordNotFoundException("Not found account id >> "+id);
 
-        return accountMapper.mapToDTO(accountRepo.findById(id).orElseThrow()) ;
+        return accountMapper.mapToDTO(entity.get()) ;
+    }
+    public Optional<Account> findByEmail(String email){
+        return accountRepo.findByEmail(email);
     }
 
 
@@ -51,6 +68,9 @@ public class AccountService {
     }
 
     public void delete(Long id) {
+        Optional<Account> entity =accountRepo.findById(id);
+        if(entity.isEmpty()&&entity==null)
+            throw new RecordNotFoundException("this account id >> "+id+" already deleted");
         accountRepo.deleteById(id);
     }
